IDEA-156648 Tree view is compacted in 2016.2 EAP
"prevScale" is used to normalize ui defaults values patched on the previous update cycle with the previous scale factor. On the first update cycle "prevScale" should be set to 1.0 as it corresponds to the scale of the ui defaults not yet patched.

(cherry picked from commit 9d4ed9ac4012bb5322273bdc3c086255cc02bc4f)