Make VCS toolwindow available for projects with <Project> mapping
The issue was introduced in 8f6589d: default mapping is a valid
mapping and shouldn't be checked here.

This fixes IDEA-161752.

Previously the availability check used ProjectLevelVcsManager.getAllActiveVcss(),
but this check became less powerful in 00adb8e: it stopped initializing
active VCSs if they weren't initialized yet.
This functionality might return back, but it will be done separately.