PY-18788 Remove redundant test as we first need to improve annotation parsing
Specifically, we need to be able to parse annotations in text form and
recognize those names that require additional import from "typing"
module. Current implementation of PyTypeParser is not capable of that,
since it omits types which names cannot be resolved in the context
surrounding the type hint.

This test was added preemptively, but it's going to take longer than
expected to fix the problem, because we want to implement the
aforementioned improvement in PyTypeParser as part of storing type
annotations in stub files (see PY-18816).