Remove ConfirmingHostnameVerifier and unused settings from CertificateManager
It turned out to be non-obvious how to properly save hostname and
expiration date exceptions in settings, so these checks were disabled
in the first place and the corresponding settings were hidden in the
configurable. CertificateManager.HOSTNAME_VERIFIER was left in
CertificateManager my mistake, and, unfortunately, has acquired some
users since then.