PY-17447 "IDE as a client" Python script debugger mode introduced
Python pydevd wrapper in the main process and in the every subprocesses being created opens server socket and waits for a single incoming connection from the IDE. These server sockets are bound to the one single port of the localhost. The port number is passed to the script in the command line. Opening server socket on the same port is achieved by using SO_REUSEPORT socket option on Linux/Mac machines and SO_REUSEADDR socket option on Windows machines. SO_REUSEPORT is supported since Linux 3.9, which is perfectly fine for docker case because docker requires Linux 3.10 and higher. A tricky connection strategy implemented in ClientModeDebuggerTransport because each debugging (sub)process requires exactly one connection but Linux could route connection to the same (sub)process twice or more. These connections are stacked on the Linux core level and IDE for a some time thinks that it is connected to a debugging process while it is actually not. Reconnecting to the same host and port solves this problem eventually.

(cherry picked from commit cccb586)