Don't ask SSH password 3 times if user pressed cancel
The ssh code allows 3 (the number is customizable) attempts to enter
a password or passphrase. However this shouldn't apply for the case
when user explicitly presses cancel in the password prompt.

To fix this we use a control-flow exception way:
throw a non-checked exception and catch only in main().

Fix for IDEA-177090