Move classes related to template languages from python-openapi back to ultimate
All their usages are in the ultimate project anyway and no third-party
plugins seem to be benefit from them (at least API Watcher claims so).
Template languages are a standard feature of PyCharm professional so
it makes little sense to keep these classes in community project.
They were migrated there in the first place since Flask plugin
belonged to the set of contrib plugins in the past.

Moreover, they contain a lot of implementation details and prevent from
getting rid of such code in python-api module as well.