Reset PythonDialectsTokenSetProvider automatically in PyTestCase#setUp()
to fix failing PythonInspectionsTest#testUnusedLocalFStringReferences()
(when it's run as part of PythonAllTestSuite).

Otherwise, the exact token sets become dependent on the order in which
tests are executed because some of them such as PyLexerTestCase
register PythonDialectsTokenSetContributors and re-initialize the
provider explicitly and then leave it in this state for the following
tests. This might cause some subtle bugs when PSI operations that
require IElementTypes other than those defined in
PythonTokenSetContributor break down all of the sudden. In particular,
this is relevant for Python fragments injected in f-strings and comments
since PyDocstringLanguageDialect uses special token type for reference
expressions.

It seems better to reset the provider automatically once and for all
than to fix it in #setUp() for every single affected test like it was
done e.g. for PyOptimizeImportsTest.