Bundled Gradle plugin overrides existing checked-in jars
The Android plugin needs builder-model on its internal classpath,
and we have a mildly out-of-date one checked in under
tools/adt/idea/android/lib.

When building the Gradle plugin at head, we need to use that
builder-model instead of the checked-in one. Otherwise, version
skew might manifest in the form of classloading errors.

Change-Id: I8850e0d0c292862acfb997217e968ca8ec4222cf