Fix property which customizes diff memory allocation
idea.max.vcs.loaded.size.kb is used to restrict memory allocations
when calculating a diff, to avoid possible OOM when comparing large
files.

However, previously it was not really possible to relax the restriction,
since diff took the minimum of idea.max.vcs.loaded.size.kb and
idea.max.intellisense.filesize (which is used as a default).

The fix is to use idea.max.vcs.loaded.size.kb is it is set, without
looking into idea.max.intellisense.filesize.

This fixes IDEA-114799