Add restored sdk root virtual file instead of original file
We check for the restored root file being contained in the excluded
paths and then we used to put the original file instead of the restored
one. For example, the restored file a Jar FS for *.egg archive was
ignored and the original file was used instead. This resulted in the
wrong assumption that our paths had been modified.

(cherry picked from commit 87ccb25)