Bugfix for migration to Stream API for sum() of List<Integer>
This commit is a patch for an incorrect refactoring of a foreach loop over a List<Integer> to a sum() call on a Stream.
When the loop variable's type is a primative int rather than an Integer, the current implementation did not add a mapToInt() call to obtain a specialized IntStream object first.

Example:
--------

List<Integer> values = ...
int sum = 0;
for(int i : values) {
  sum += i;
}

The code above would refactor to:
int sum = values.stream().sum();

The correct refactoring is:
int sum = values.stream().mapToInt(i -> i).sum();