Prevent double disposal of welcome frame
The welcome frame extends JFrame and implements
com.intellij.openapi.Disposable. This means that:
1) It overrides dispose() and must call super.dispose()
to work correctly
2) Methods which can accept an instance of it by a JFrame reference
parameter and intend to dispose it, must check whether the JFrame
is actually an IntelliJ Disposable. If yes, then dispose() is not
intended to be called directly - instead, Disposer.dispose()
should be used. If it's a regular JFrame not implementing Disposable,
then the original JFrame.dispose() can be called directly.

Amongst all, this approach prevents multiple disposals of the same object.
Effectively, every class extending JFrame and implementing Disposable
is susceptible to the double disposal sorts of bugs. A quick search
over the code base didn't suggest any other immediate candidates though,
but it's worth keeping an eye on this (or use a proper inheritance
tree analyzer).

Bug: http://b.android.com/209877
Change-Id: I7b1d7b0a890544bec9dbd4e2c264d726147589e2