PY-21244 Add a new parser for PEP 3101 new-style formatting syntax
The new parser handles nested replacement fields in the "format spec"
part of a field, can deal with incomplete/unterminated strings and
fields and, in general, is capable  of extracting more information
about them such as ranges of all field parts, its attributes and
lookups, etc.

It's intended to replace much more limited PyStringFormatParser
and to be reused in the dedicated intention about new-style formatting
(PY-8325) and resolving references injected inside string literals used
with .format() method. For now, to ease migration it was retrofitted,
so that returned Field objects are compatible with
PySubstitutionChunkReference.