Fixing SearchScope for highlight duplicate classes
When we have two different classes with the same name, one being an
AndroidTest and the other an UnitTest (which is perfectly acceptable
since these are different scopes), the IDE highlights the class pointing
it as a duplicate. It happens because HighlightClassUtil is searching in
the entire module scope and not only in the subscope determined by the
class.

This CL fixes this problem by passing the correct search scope (the
scope determined by the class element instead of the entire module
scope).

Patch applied by intellij in master in commit:
40147bba55cfcf102fff50400894bad1de52285a

Fixing bug: http://b.android.com/208625

Change-Id: I1d668b63da128f453244a4045ca7c3af5646cc91