Add option to not log error on missing jars in JarLoader
Android Studio preview makes use of the UrlClassLoader to load
classes that are part of the project. In some cases, like when building,
some classes might be temporarily missing.
The preview handles that case correctly but the JarLoader logs an error
that shows the error notification.
This change adds an option that its used from Android Studio to disable
that behaviour in the preview.

Bug: http://b.android.com/226800
Test: Platform change to be upstream to JetBranins
Change-Id: If34368301fa283e78752d8f2ae66035592c38059