Refactoring modules to allow UI tests to run through IJ
The problem is that when running UI tests through IJ, we have to have
everything that's included in production on the classpath. This can
exclude test-only modules, and, in the case that that module has a
plugin.xml that replicates some of the plugin.xml in android-plugin, it
must be excluded.
When running tests through ant, test modules must be included
(obviously, since we need to run the tests). But we can prune those out
via the gant script.

This change creates a new module, community-main-tests, which includes
community-main and those test modules that have their own plugin.xml.
This module can then be used for running the ant tests, while
android-uitests can use the original community-main.

Test: This should allow existing UI tests to run through IJ. No
additional tests are needed/possible.

Change-Id: Iaf460812b58119fcf42f8c5ac213e3c64a5e120b