[vcs-log] change RefGroup so that is returned several colors
This should help with proper HEAD label display in the table: even for expanded groups, HEAD reference should not be shown in the list, yet the group still has to have a yellow tag.
Also, SimpleRefGroup class from Mercurial plugin is extracted and moved to the vcs-log impl. This class now can be used from git and mercurial, and it has a proper getColors method moved from LabelPainter.