Avoid dependencies on jna-platform.jar
The launcher runs with "jna.jar" in the classpath, but not with
"jna-platform.jar".

Change 2efe732bad524a846ac36196ddcbdfaca431c3d0 introduced a
dependency on jna-platform (Windef.DWORD and WinDef.HMODULE).

It is easier/safer to avoid the dependency than to update the
launcher script to include an additional jar. This is what this
change does.

BUG=https://code.google.com/p/android/issues/detail?id=228003

Test: Manual
Change-Id: Ibf127e0c50742fd82b679691f1d93c7168fcf409