Fixing symlink check again.
There was a problem where a file was considered a symlink if there was a
symlink anywhere in the path.
Since JB now uses Java 1.8 in updater and we've decided to switch over
also, replaced the custom symlink code with appropriate Files calls.
This definitely fixes one issue mentioned in b.android.com/227863, and
may well fix the originally reported issue as well (though it isn't
fully understood).

Test: Upated PatchFileCreatorTest.

Change-Id: I098eea1b2634563876c8156377f4d6834e1d4409