[vcs-log] index commits one by one
During indexing when only a part of the repository is being processed, commits were loaded in batches by 1000, then each batch was analysed and stored in the index. After that commits themselves were (and still are) thrown away. Since it turned out commits can be quite big, it is unwise to collect them. So now commits are processed immediately as they are loaded.

IDEA-164274