Don't show android-tests and designer-tests at the top level for android
The android-tests and designer-tests modules are actually present
entirely within the android and designer modules respectively. But by
grouping them within the "Android" group, they show up at the top level
(next to android and designer). But that structure is confusing because
the IJ navigator displays it as "testSrc [android-tests]" where it has
the folder name of the module first, which happens to be testSrc for
both these modules.

This CL simply removes them from the android group, so they just show up
as modules inside the android and designer modules respectively.

Test: N/A
Change-Id: Iec162e4f42f42b77e2fd351384956ac7844f48d1