Don't allow duplicate VCS roots
Use a MultiMap backed by an OrderedSet, this not only avoids the dups,
but simplifies the code.

The set is ordered to keep the original order written in the vcs.xml,
to keep ability to use this order in future (e.g. process dependent
roots from the base to the client).