Don't report manually unregistered roots
If a user manually removed a root from the mappings, no need to report
this root as unregistered, since it was an intentional change.

On the other hand, if such root was added back, it should be removed
from ignored just in case (e.g. if vcs.xml get accidentally updated).