Automatically add a single Git/hg root
If we detected that there is a single unregistered root in the project,
which wasn't reported previously, and there are no other roots,
add this root automatically.

If user doesn't want VCS integration, he can easily remove the root
from the Version Control preferences, and it won't be reported or
auto-added anymore.

Add a registry option for a while (on by default)
in case something goes wrong with this behavior.

This fixes IDEA-118475