[ui] fix fade-out in loading decorator with smooth scrolling enabled
In order to have smooth scrolling, loading layer should not be on top of viewport, but rather added on demand.
But since loading layer has fade-out, it can not be removed immediately when setLoadingLayerVisible is called.
It should be removed only at the end of fade out.
Before that removal, it should paint with zero alpha, in order not to leave artifacts after itself.