git: collect affected files from correct stdout/err stream
When the message is detected (e.g.
"untracked working tree files would be overwritten by merge"),
remember the output type, and read the files only from the stream
of that type.

Before that, since stdout and stderr are reported asynchronously,
messages could arrive like that:
    error: The following untracked working tree files would be overwritten by merge: 
    Updating 48295bf..cfba4dc 
    untracked.txt
And we treated "Updating..." as a file.

This caused blinking of the GitBranchWorkerTest.

Also add some debug logging for possible future needs.