Don't show function parameters in the list of suggestions for import
Otherwise it becomes difficult to determine the right signature to
display since there might be several overloads of the same function
in .pyi files.

Overall, if there were several definitions of the same function, for
example, in the actual module and in the .pyi file/user skeletons, it
wasn't clear which signature would be displayed because it depended on
the order in which functions were returned from the index in
PythonImportUtils.addSymbolImportCandidates()
(it was discovered thanks to a blinking test on TC).