git: fix race condition in reading git command output
stdout and stderr are read by different threads, each of which modify same output collections.
These collections are non-thread-safe ArrayLists.

Because of this lack of synchronization, the list could get null elements,
which is unexpected by other code. null elements can appear if the size of the backing array and its elements are updated without proper synchronization.