IDEA-166747 Freeze after calling duplicate action repeatedly
evaluateLanguageInRange can be called with a range from the editor with
uncommitted document. There's no point in repeatedly doing findElementAt for offsets outside PsiFile. It should be enough to just traverse the hierarchy up until an element that covers the selection or the file, and return the language of the found element