[vcs-log] try to fix a flaky test
In the test it is checked that the first log read reads only first few commits (in order to show them quickly). But also readFirstBlock sends RELOAD_ALL request, which is processed immediately, since we call readFirstBlock outside of SingleTaskController. It seems that sometimes we are able to reload in background and set a new (full) data pack before returning from readFirstBlock method. Because of this full data pack is returned. It would not be a problem in production (it is good to return all the data immediately) but since we want to do a proper test here, we block full log loading temporarily.