ui: stabilize action position in Keymap tree
This fix affects actions without default text/description
set in a plugin.xml or ActionsBundle.

Text/description might be initialised in AnAction constructor,
and thus be empty in StubAction.
Some ActionGroup might contain stub action, even if it was already loaded.

While collecting actions for "Other" group we will load non-stub version
of an action, and it will pass the filter.

Thus, prior to this fix, AnAction might change its place in a tree
depending on whether text filter was used or not.
And, in some cases, depending on the filter itself - as it could've match
internal action id.