Adds new macro expansion $MODULE_CONTENT_ROOT$
This new macro is expanded to the current module's content root when a
the module only has one content root.
I have never seen a case were a module has multiple content roots, and
am therefore unable to decide what to do in these cases.
When a project uses the file based project descriptor (*.iml), this
should be strictly equivalent to $MODULE_DIR$
When a project uses the directory based project decription (.idea/*),
this macro points to the content root instead of the parent directory of
the module file.

I am unfamiliar with the codebase and was unable to find a test which
exercises the code path I changed. I tried writing one based on
`PathMacroManagerTest` but was unable to make it work as it seems to
requires a lot of setup which goes beyond the setup already provided.

However please consider incuding this since it opens a way to fixing:

https://youtrack.jetbrains.com/issue/IDEA-152227
https://youtrack.jetbrains.com/issue/SCL-9296
https://youtrack.jetbrains.com/issue/IDEA-121819
https://youtrack.jetbrains.com/issue/IDEA-153852
And probably a few more $MODULE_DIR$ related issues