Fix project leaks in VCS tests
Nullize fields before calling super.tearDown(), because the latter
analyzes the test for leaks.

clearFields intentionally doesn't affect the PlatformTestCase,
so the project will be correctly disposed in super.tearDown().

dumpTogToStdOut should be called before clearFields, otherwise
the myTestStartedIndicator gets null, and no output is dumped.
In this case we won't get this dump if super.tearDown() fails,
but it is not a big problem, since the whole DEBUG log is available
anyway.