[file-history] do not create vcs file revision for deleted file, rewrite show all affected paths
We do not have a file revision for deleted file (and can not get or annotate it), so do not create revision for it. One action that was affected by this change was show all affected paths. But turns out this action does not need revision at all, having full details it is easy to show changes browser for the commit.