Adjust code-style settings for imports
… based on those in the platform/tools/base IDEA project, which includes
many of the same source files. This way the overlapping projects don't
clash due to, for example, different sort orders.

More important, this avoids on-demand imports for reasons articulated at
http://errorprone.info/bugpattern/WildcardImport
and because replacing individual imports with wildcards, and vice versa,
is a continual source of churn caused by fluctuations in the number of
classes imported from a single package crossing the arbitrary threshold.

Note: This is intended for Android Studio code. For the rare times we
modify code in platform/tools/idea, we should continue to avoid
unnecessarily changed lines, including imports.

Change-Id: Idc1af15de13a27ab3b45f90e4b110c82d35cc6f8