[git] set diff.renameLimit to zero by default when invoking git log command
This commit fixes IDEA-123397, which was broken after introducing new file history based on log with indexes.

Diff.renameLimit is the maximal size of commit for which git performs rename detection. Zero value means no restrictions. This allows to detect renames even in big commits during indexing, that provides data for building corret file history for files changed in this commits.

Since git process may take a lot of memory for detecting renames in big commits, a registry option is introduced to switch it off.