Don't take the HeavyProcessLatch for VCS operations
It was introduced to let Automake functionality suspend itself, while
some Git operations are changing working tree, to avoid compiling
the project in "intermediate" (e.g. partially checked out) state.

However, currently it is enough to send an event to the
BatchFileChangeListener, which is listened by Automake, so the latch
can be removed.

On the other hand, taking the HPL suspends the highlighting daemon,
which is unfortunate since doesn't allow user to see analysis
results until the VCS operation completes, and leads to issues like
IDEA-160945.