Restore binary compatibility
This is a follow-up to 3cdd5e9.
It appeared that there are 3-party usages of the methods,
so the method signature change "AccessToken -> void" breaks them.

Another approach would be to deprecate both methods and rename
workingTreeChangeStarted to something different, but it looks like
the AccessToken approach look fine as well, and doesn't require
introducing new methods.

Relates to IDEA-160945.