Fix issue with disposed project in ProgressWindow
When closing a project, IntelliJ sometimes starts background processes
like indexing that are called inside invokeLater from
DumbServiceImpl.queueTaskOnEdt. When those processes are finally run,
they try to create a ProgressWindow which wants to be registered as a
child of the project. However the project may have been disposed in the
meantime, which triggers an exception.
To reproduce, create a new project with C++ support, wait for it to be
fully loaded, and close it.

Test: Many UI tests, such as NewProjectTest
Change-Id: I7db2d322eaa9592d84b41e2396895a012c744920