Don't autodelete directory after deleting the last file in it
Executing `git rm` on the last file in a directory makes Git remove not
only this file, but its parent directory as well.

Although it is Git's behavior, it is not convenient, and it is also
unexpected since `git rm` is not explicitly triggered by user who
deletes a file from the Project View.

GitVFSListener is triggered _after_ actual deletion of the file,
so the only thing Git should do is to delete the file from the index:
git rm --cached. This does the same as `git rm`, but prevents Git from
removing the directory (since files are not touched on disk, only
in the index).

Fixes IDEA-75590

(cherry picked from commit dcac402)