[vcs-log] when filters change, recalculate columns size
When filters change, commits with author with longer names can become visible. This often leads to author column being too small.
So need to recalculate preferred size in this case.

This of course creates a new problem: if user resizes a column, it should not be automatically adjusted. This is going to be fixed in later commits.

(cherry picked from commit 0a0c40c)