[vcs-log] save date and author columns width IDEA-168370
Minimum width of the columns is now calculated from data, preferred -- from user settings. This does not allow to set width less then data length.

Since we want fallback to auto-resized columns possible, when columns are resized to the minimum it is assumed that they need to be auto-resized from now on. -1 is saved to the settings in this case.

(cherry picked from commit 2445400)