Allow to remove Git root mapping
The bug was introduced in be3de925 where Git roots became auto-added
in simple cases.

The IGNORED_UNREGISTERED_ROOTS setting was read only once - on the
service initialization, i.e. on startup. The fix is to take it into
account every time.

On the other hand, there is no need to pre-populate
myReportedUnregisteredRoots with ignored unregistered roots, since
it is a different conception: reported roots are not explicitly ignored
by the user, they are stored to be reported once per session.

The reason why this bug was not noticed by the author of the original
change is unclear, but probably the root was reported before testing
the fix and became remembered in the field.

Fixes IDEA-168690.