Avoid deregistering roots that will registered again
Previously, once a new set of roots was received,
all the current roots were first deregistered and
then the new ones registered.

In case the majority of the roots belonged to the both
lists, this caused two issuses:

  - Performance overhead

  - Loss of event if it happened after deregistering
    before registering

To avoid those issues, the set of current and new roots
will be analyzed and only new roots are registered and
only the old roots no longer present are de-registered.