standalone JPS distribution: extract 'java-runtime' module to a separate JAR
This is needed to ensure that the whole jps-builders.jar won't be added to classpath when javac process is forked. Otherwise the classpath will include *.class files which aren't compatible with Java 6, and it'll lead to exceptions if javac from JDK 1.6 is used for compilation.

(cherry picked from commit 893f8a2, IDEA-CR-18894)