[git] use double record start and end symbols IDEA-156249
Symbols \u0001 and \u0003 were used as markers in order to parse git log command output. Problem is, these symbols also can occur in commit messages (amazingly enough). This led to exceptions during loading of the log, that prevented users to see it (despite some measures against it). It is unclear how to eliminate the problem completely, but it seems to be possible to reduce its likelihood by doubling these separator symbols.