Add a way to skip git pre-commit and pre-push hooks (IDEA-115914)
Git supports a `pre-commit` and `pre-push` hooks that are executed
after a commit or a push accordingly (https://git-scm.com/docs/githooks).
Sometimes it's needed to skip the hooks and bypass the checks for them.
For that purpose git supports a parameter `no-verify`.
With that commit inside the git commit dialog a checkbox "Skip hook" will be
available if any of the repositories have a `pre-commit` hook installed.
Similar way in the push dialog it will be possible to bypass the
`pre-push` hook.
Options are not visible and not used when there's no such hooks installed.
Checkbox options are not saved for the next time because that may
lead accidentally to commits or pushes without proper hook verification.

Currently only local hooks (inside .git/hooks folder) are supported.
It doesn't support a new git configuration variable `core.hooksPath`
that allows customizing where the hook directory is (since git 2.9.0).