Allow to remove <Project> mapping
It used to be added to ignored roots as is, i.e. as empty String,
while roots are detected as real paths.

The fix is to ignore it by "project base path" which is not very
well-defined, but works fine in most cases.

Follow-up to e736a37
Fix for IDEA-168993

(cherry picked from commit 6b47cdba94d8c8eb0bf01c80466790e5e1c0f595)