Allow to remove <Project> mapping
It used to be added to ignored roots as is, i.e. as empty String,
while roots are detected as real paths.

The fix is to ignore it by "project base path" which is not very
well-defined, but works fine in most cases.

Follow-up to e736a37
Fix for IDEA-168993