PY-21180 Avoid possible collisions of caches in PyUtil#getParameterizedCachedValue()
First, each cache has the unique user data key, associated with the
function passed to getParameterizedCachedValue(), as opposed to
the instance CachedValueProvider which is actually the same for all
caches created this way.

Second, in QualifiedNameFinder#findCanonicalImportPath() use pair of
(symbol, anchor) as the cache key so that the lambda doesn't capture
values from outer scope.

(cherry picked from commit 7665179)