Disable the "new in 1.8" inspection: redundant in Android Studio
IntelliJ has a basic API level check which looks for things like
usages of the new streams API which requires 1.8. However, this
is tied to the JDK, not the Android SDK, and is unaware of things
like the minSdkVersion -- so disable it in Android Studio. It's
completely redundant in Studio anyway since we have lint's API
check which does the same thing but correctly for Android.

Test: This just turns off functionality.

Change-Id: I8b9a7822cd435048f4477ef3b46599c3c6cb9473