Switch severity of SuspiciousMethodCalls to error
These issues are almost always real bugs, and we had a violation
in our code. The new version of bazel and errorprone already breaks
the build when it finds this, so we might as well highlight it for
developers as soon as they touch the code.

(The other edits to the inspection profile were made automatically
by IntelliJ 2017.1.)

Test: Just a project metadata change, no production build impact

Change-Id: Ib486de5b0c767e332f11508ba86d9797a3a11704