Force using IntelliJ file chooser for local Python SDKs due to symlinks and hidden files (PY-21787)
On macOS the native file chooser follows symlinks making it impossible
to select virtualenv interpreters, meaning the base interpreter is
selected instead of the virtualenv interpreter.

Native macOS dialogs also make it difficult to select virtualenvs from
hidden folders, especially from ~/.virtualenvs/ which is quite common.
You have to know a special keyboard shortcut to do it.

We are switching to the IntelliJ dialog until those problems are
addressed in the native macOS file choosing dialog.