Refresh only that part of the VFS that has been changed by Git operations
Refreshing the whole VFS is reliable but slow, especially in case of feature branches, when branches diverge just a little, and refreshing only a couple of files is required which is much faster.

On the other hand, collecting the difference between current working tree and the target reference is quick enough: even `git diff origin/143` takes ~1–2s.

A registry option is introduced to be able to turn the optimization off in case there is a report that not files were refreshed.

Following operations are optimized:
* reset (including "undo commit"): uses `git diff <revision>` which compares with working tree: working tree may change as the result of reset --hard.
* checkout and merge: use `git diff HEAD..<branch>`.
  Comparing with working tree is unnecessary ("local changes would be overwritten by checkout/merge" means that local changes are not touched by these operations).
  Merge should actually compare with the merge base (`git diff HEAD...<branch>`), but that's much slower, while refreshing several more files is not that slower.
* rebase can happen together with checkout (`git rebase master topic`), so at first it checks the diff between current branch and the branch which is to be checked out; and then it additionally checks the diff for the rebase itself (again, should be calculated with "..." for more precise results, but ".." is used because it is faster).

Note that the reason ".." can be used instead of "..." is because we don't care about the changes themselves and how accurate they are: we just need a superset of affected files.

Only trivial cases are covered: e.g. smart checkout, rollbacks, conflicts, continue rebase, etc. still perform full refresh for now, since they are relatively rare and a bit trickier).

This fixes IDEA-163043