Remove unused action name/description declaration
It is actually used in one 3-party plugin: Changes Bar.

But all builds of the plugin are very old and incompatible with
the current version of IDEA. Only the first version is compatible
(until-build is not defined), but it breaks anyway with other errors.

Follow-up to 69df742
Relates to IDEA-66052