PY-23125 Preserve blank lines before the first comment in optimize imports
I tweaked Python import optimizer so that it inserts new imports
and their comments after the last of original imports, not before the
first of them (and then deletes the old ones). The reason is that it
turned out to be practically impossible to use PsiComment as the anchor
for PsiElement#addRangeBefore(), because of the hardcoded behavior of
CodeEditUtil.addChildren() which skips preceding whitespaces in this
scenario.

Otherwise, if the blank line before "# noinspection" comment preceding
the first imports statement gets removed, next time this comment will be
left in place, considered part of the top-level module comment
e.g. license.