Make google-api-java-client a module library
Project libraries get packaged under lib/, and are loaded by the IDE's
core classloader. The google-api-java-client library is only needed by
two plugins, google-login-as and google-cloud-tools-core-as. Moreover,
it repackages some of its dependencies (Guava is jarjar'ed in there) but
not all -- it also depends on javax.servlet-api, which means they need to
be in the same classloader.

With this CL, both google-api-java-client and javax.servlet-api should be
loaded by the plugin classloader of google-login-as, and available only to
downstream dependents.

Bug: http://b.android.com/235111
Bug: http://b.android.com/253707
Test: N/A
Change-Id: Ibc20c6d6a46208f956c5beb2890b9a16fee5b77d