[git] skip grafted in commit references list
When git repository has grafted commits, or is a shallow clone, commits with rewritten parents have "grafted" marker. It is shown with the list of references for the commit in log, but this is not a reference. Since creating a "grafted" reference leads to exceptions (having several references with the same name but a different hash is rather unexpected), this commit skips "grafted" markers entirely.

IDEA-169831, EA-99213

(cherry picked from commit df80a4a)