Add exception registry
This CL adds an "exception registry", an object which keeps track of
all the exceptions that it has been told about.  These are stored in
an efficient manner, such that if you for example get 5000 reports of
a particular exception, that takes the same amount of storage as just
1.

It does	this by	splitting a stacktrace up into its individual frames,
and storing these frames in a tree along with frequency	counts.	When a
new stacktrace comes in, each	successive frame is matched with
existing frames	in the tree and	the counts increased, and for new
parts of a frame new nodes are inserted	into the tree.

With this data in place	the ExceptionRegistry can provide answers
about the exceptions that have been reported - it can for example
return the most	frequently reported exception, or return all
exceptions, or return a	"report", a string which has a table of
hashcodes for the reported exceptions sorted in	descending frequency
order, optionally with a collapsed stacktrace summary on the right
such as

11 379B89010804F9BC03263B09AF393F22 FileNotFoundException: FileIn…
 2 5E389B3192074C7A8260CA82B0ECEE09 XmlPullParserException: KXmlP…
 1 6F4781E61570E16948ECCE200C16B548 NullPointerException: DateUti…
 1 E287C482DD9EA73DD035362F503A7183 ArithmeticException: MyCustom…
 1 ADDCF4E7B9F2324965756AAE9CDD524C FileNotFoundException: Compon…
 1 759507990F4C30449A488B4538B92EB5 InternalException: JDWPExcept…

The above is truncated due to line width in git	commit messages	but
the registry allows you	to specify arbitrary width (default is 300)
and you	get something like

ComponentManagerImpl.init:104←ApplicationImpl.load:425←411←IdeaApp...

The next step is to hook this up to the	SystemHealthMonitor such that
we can (a) send	MD5 checksums with crash reports, and (b) send the
crash frequency table occasionally.

Test: Unit test included

Change-Id: I0f19b2191311145ab147c82eeac3c184cf12eaec