updater-ui was incorrectly included in studio build
After the build script rewrite, the updater-ui module was incorrectly
getting included in studio. This meant that in PatchRunner where we try
to load RepoUpdaterUI from the patcher jar, the classloader was getting
confused because RepoUpdaterUI was available in the parent classloader
(the normal studio one), but the superclass (SwingUpdaterUI) was not,
since it's in the "updater" module that was correctly included in the
patcher and not in studio.
This change removes updater-ui from studio.

Fixes b.android.com/255171

Test: this problem only occurs on a full build of studio, so it seems
like the only real place to do the testing would be in a ui test. But we
don't really have the capability of running sdk package installs in ui
tests, so...

Change-Id: If9fe9db0f7ca276c503b3bab61cd5d6ec2b06bee