Track more details along with exception metrics
Earlier, we only tracked the count of exceptions and crashes via
the metrics pipeline. The exception stack trace was uploaded, but it
was throttled, which meant that it was possible that we could lose
some information.

This CL uses the recently added ExceptionRegistry to improve our
diagnosis of these frequent exceptions:
 - We send the MD5 and summary stacktrace of all seen exceptions
   via the metrics pipeline.
 - We also include the same MD5 and summary in the full exception
   report sent to Crash.

Together, this should allow us to have a complete picture of all
the exceptions that could happen.

Bug: http://b.android.com/237687
Test: manual
Change-Id: I60fcb088adcfec89efddd1a1a844a6a9d5f533d8